refactor: 공용 텍스트인풋 컴포넌트 분리
