+++ frontend/app/src/main/java/com/apptive/japkor/ui/components/CustomTextField.kt
@@ -1,0 +1,39 @@
+package com.apptive.japkor.ui.components
+
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.apptive.japkor.ui.theme.CustomColor
+
+@Composable
+fun CustomOutlinedTextField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    placeholder: String
+) {
+    OutlinedTextField(
+        value = value,
+        onValueChange = onValueChange,
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(50.dp)
+            .border(
+                width = 1.dp,
+                color = CustomColor.gray100,
+                shape = RoundedCornerShape(10.dp)
+            ),
+        shape = RoundedCornerShape(10.dp),
+        placeholder = {
+            CustomText(
+                text = placeholder,
+                type = CustomTextType.bodyMedium,
+                color = CustomColor.gray300
+            )
+        }
+    )
+}
